+++ app/src/main/java/com/example/playlistmaker/SearchActivity.kt
@@ -19,6 +19,7 @@ import androidx.recyclerview.widget.RecyclerView.Recycler
 class SearchActivity : AppCompatActivity() {
 
     private var restoredText = ""
+    private val trackList = TrackList()
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -71,22 +72,9 @@ class SearchActivity : AppCompatActivity() {
         inputText.addTextChangedListener(searchTextWatcher)
 
         val recycler = findViewById<RecyclerView>(R.id.search_list)
-        val tracks = listOf(
-            Track(
-                "Smells Like Teen Spirit",
-                "Nirvana",
-                "5:01",
-                "https://is5-ssl.mzstatic.com/image/thumb/Music115/v4/7b/58/c2/7b58c21a-2b51-2bb2-e59a-9bb9b96ad8c3/00602567924166.rgb.jpg/100x100bb.jpg"
-            ),
-            Track(
-                "Billie Jean",
-                "Michael Jackson",
-                "4:35",
-                "https://is5-ssl.mzstatic.com/image/thumb/Music125/v4/3d/9d/38/3d9d3811-71f0-3a0e-1ada-3004e56ff852/827969428726.jpg/100x100bb.jpg"
-            )
-        )
+
         recycler.layoutManager = LinearLayoutManager(this)
-        recycler.adapter = TrackAdapter(tracks)
+        recycler.adapter = TrackAdapter(trackList.tracks)
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
