+++ app/src/main/res/layout/track_list_item_layout.xml
@@ -40,8 +40,9 @@
             <TextView
                 android:id="@+id/track_artist"
                 style="@style/TracksArtists"
-                android:layout_width="wrap_content"
+                android:layout_width="0dp"
                 android:layout_height="wrap_content"
+                android:layout_weight="1"
                 tools:text="trackArtist" />
 
             <ImageView
