+++ app/src/main/res/values/styles.xml
@@ -4,6 +4,7 @@
     <style name="BG_color">
         <item name="android:background">@color/YP_blue</item>
     </style>
+
     <style name="BG_icons">
         <item name="android:tint">@color/YP_black</item>
     </style>
@@ -43,6 +44,7 @@
 
     <style name="TracksNames">
         <item name="android:textColor">@color/YP_black</item>
+        <item name="android:singleLine">true</item>
         <item name="android:textSize">16sp</item>
         <item name="android:gravity">center_vertical</item>
         <item name="android:textFontWeight">400</item>
@@ -50,6 +52,7 @@
     </style>
 
     <style name="TracksArtists">
+        <item name="android:singleLine">true</item>
         <item name="android:textColor">@color/YP_light_gray_text</item>
         <item name="android:textSize">11sp</item>
         <item name="android:gravity">center_vertical</item>
